requirements.txt
google-cloud-bigquery==3.10.0
requests==2.31.0
